+++ frontend/fingertips-frontend/components/organisms/Inequalities/inequalitiesHelpers.test.ts
@@ -5,10 +5,13 @@ import {
   groupHealthDataByInequalities,
   groupHealthDataByYear,
   Inequalities,
+  InequalitiesTableRowData,
+  mapToInequalitiesTableData,
   shouldDisplayInequalities,
 } from './inequalitiesHelpers';
+import { GROUPED_YEAR_DATA } from '@/lib/tableHelpers/mocks';
 
-const MOCK_HEALTH_DATA: HealthDataForArea = {
+const MOCK_INEQUALITIES_DATA: HealthDataForArea = {
   areaCode: 'A1425',
   areaName: 'North FooBar',
   healthData: [
@@ -53,12 +56,12 @@ const MOCK_HEALTH_DATA: HealthDataForArea = {
 
 const yearlyHealthDataGroupedBySex = {
   2004: {
-    All: [MOCK_HEALTH_DATA.healthData[2]],
-    Female: [MOCK_HEALTH_DATA.healthData[3]],
+    All: [MOCK_INEQUALITIES_DATA.healthData[2]],
+    Female: [MOCK_INEQUALITIES_DATA.healthData[3]],
   },
   2006: {
-    All: [MOCK_HEALTH_DATA.healthData[0]],
-    Male: [MOCK_HEALTH_DATA.healthData[1]],
+    All: [MOCK_INEQUALITIES_DATA.healthData[0]],
+    Male: [MOCK_INEQUALITIES_DATA.healthData[1]],
   },
 };
 
@@ -83,11 +86,17 @@ describe('should display inequalities', () => {
 describe('groupHealthDataByYear', () => {
   it('should group health data by year', () => {
     const healthDataGroupedByYear = {
-      2004: [MOCK_HEALTH_DATA.healthData[2], MOCK_HEALTH_DATA.healthData[3]],
-      2006: [MOCK_HEALTH_DATA.healthData[0], MOCK_HEALTH_DATA.healthData[1]],
+      2004: [
+        MOCK_INEQUALITIES_DATA.healthData[2],
+        MOCK_INEQUALITIES_DATA.healthData[3],
+      ],
+      2006: [
+        MOCK_INEQUALITIES_DATA.healthData[0],
+        MOCK_INEQUALITIES_DATA.healthData[1],
+      ],
     };
 
-    expect(groupHealthDataByYear(MOCK_HEALTH_DATA.healthData)).toEqual(
+    expect(groupHealthDataByYear(MOCK_INEQUALITIES_DATA.healthData)).toEqual(
       healthDataGroupedByYear
     );
   });
@@ -96,14 +105,17 @@ describe('groupHealthDataByYear', () => {
 describe('groupHealthDataByInequalities', () => {
   it('should group health data by sex', () => {
     const healthDataGroupedBySex = {
-      All: [MOCK_HEALTH_DATA.healthData[0], MOCK_HEALTH_DATA.healthData[2]],
-      Male: [MOCK_HEALTH_DATA.healthData[1]],
-      Female: [MOCK_HEALTH_DATA.healthData[3]],
+      All: [
+        MOCK_INEQUALITIES_DATA.healthData[0],
+        MOCK_INEQUALITIES_DATA.healthData[2],
+      ],
+      Male: [MOCK_INEQUALITIES_DATA.healthData[1]],
+      Female: [MOCK_INEQUALITIES_DATA.healthData[3]],
     };
 
-    expect(groupHealthDataByInequalities(MOCK_HEALTH_DATA.healthData)).toEqual(
-      healthDataGroupedBySex
-    );
+    expect(
+      groupHealthDataByInequalities(MOCK_INEQUALITIES_DATA.healthData)
+    ).toEqual(healthDataGroupedBySex);
   });
 });
 
@@ -111,7 +123,7 @@ describe('getYearDataGroupedByInequalities', () => {
   it('should group year data by sex', () => {
     expect(
       getYearDataGroupedByInequalities(
-        groupHealthDataByYear(MOCK_HEALTH_DATA.healthData)
+        groupHealthDataByYear(MOCK_INEQUALITIES_DATA.healthData)
       )
     ).toEqual(yearlyHealthDataGroupedBySex);
   });
@@ -132,5 +144,56 @@ describe('getDynamicKeys', () => {
     expect(
       getDynamicKeys(yearlyHealthDataGroupedBySex, Inequalities.Deprivation)
     ).toEqual(expectedKeys);
+  });
+});
+
+describe('mapToInequalitiesTableData', () => {
+  it('should map to inequalitiesSexTable row data', () => {
+    const expectedInequalitiesSexTableRow: InequalitiesTableRowData[] = [
+      {
+        period: 2004,
+        inequalities: {
+          Male: {
+            value: 703.420759,
+            count: 267,
+            upper: 578.32766,
+            lower: 441.69151,
+          },
+          Female: {
+            value: 703.420759,
+            count: 267,
+            upper: 578.32766,
+            lower: 441.69151,
+          },
+        },
+      },
+      {
+        period: 2008,
+        inequalities: {
+          Persons: {
+            value: 135.149304,
+            count: 222,
+            upper: 578.32766,
+            lower: 441.69151,
+          },
+          Male: {
+            value: 890.328253,
+            count: 131,
+            upper: 578.32766,
+            lower: 441.69151,
+          },
+          Female: {
+            value: 890.328253,
+            count: 131,
+            upper: 578.32766,
+            lower: 441.69151,
+          },
+        },
+      },
+    ];
+
+    expect(mapToInequalitiesTableData(GROUPED_YEAR_DATA)).toEqual(
+      expectedInequalitiesSexTableRow
+    );
   });
 });
